Test existance of default CSS before reading it
If no CSS files are specified we attempt to load in the default
stylesheet, which defaults to "/stylesheets/email.css". Since this file
may or may not exist in every Rails app, its existance should be
verified before attempting to read it.

This dies silently if the file does not actually exist, but when running
mailer specs it clutters the output significantly as File continually
warns you about the nonexistance of email.css