Add support for attachments
Or better, just don't mess with the attachments. The solution to the
problem is the following:

If no text generation is required (disabled or text part present):
- replace the html part with the processed html part

If text generation is required:
- create a multipart/alternative that containing an html part with the
  processed html and a text part containing the generated text
- replace the html part with the multipart/alternative part

Special care needs to be taken when the email contains only the root
part (the message itself), then the properties of the newly generated
part (html part or multi part) need to be copied over to the message.