Prefer precompiled assets over the pipeline
This alert the order of loaders so that the
assets are loaded from file system first and
only then, if failed, try to use the pipeline.

This should solve the problems with Rails 4
which always has the assets pipeline enabled
and thus always uses pipeline to precompile assets
instead of picking up ready-to go files.

This commit will also resolve the issue for those
who choose not to run full `assets` group in production
to save tons of memory.
(More at
https://github.com/rails/rails/commit/49c4af43ec5819d8f5c1a91f9b84296c927ce6e7#commitcomment-4544418)