Make NetworkLoader more resilient.
If the Rails asset host is set without a scheme (which is the
recommended way to set an asset host so that it works both with HTTP and
HTTPS pages) the `NetworkLoader` will fail because it requires a scheme to
be present. This change allows the `NetworkLoader` to work with asset
hosts that do and don't include the http scheme.

This change also removes the dynamic setting of `scheme` and `host` on
URI which does not work (the GET call still fails since `request_uri` is
undefined).

I've included tests to make sure the `NetworkLoader` now works with the
asset host set (with and without a scheme) and does not cause exceptions.