Fix encoding
After extensive testing, there should be no need to mess with the
transfer encoding or the body encoding. All we need to ensure is that
the proper charset is configured in the `Content-Type` header when
assigning the body, which needs to be the encoding of the string being
set as the body.

Fixes #240